refactor(ooda_actions/goal_session): extract inline tests to siblings
goal_session.rs (1594 → 862). 4 inline test mods extracted to:
- tests_goal_session_inline.rs (511) — main tests mod
- tests_goal_session_validators.rs (220) — label/placeholder/makework

Promoted to pub(super): assess_only_outcome, is_plausible_label,
action_is_valid, is_placeholder_echo, is_makework_title.

All 75 goal_session tests pass. Source split (advance_goal_with_session
+ gh helpers) follows in next PR.

Refs #1266
