+++ src/ooda_actions/mod.rs
@@ -19,6 +19,10 @@ mod tests_dispatch;
 mod tests_goal_session;
 #[cfg(test)]
 mod tests_goal_session_extra;
+#[cfg(test)]
+mod tests_goal_session_inline;
+#[cfg(test)]
+mod tests_goal_session_validators;
 
 use crate::error::SimardResult;
 use crate::ooda_loop::{ActionKind, ActionOutcome, OodaBridges, OodaState, PlannedAction};
